+++ frame/src/main/java/com/dreampany/frame/data/source/room/StoreRoomDataSource.java
@@ -64,12 +64,12 @@ public class StoreRoomDataSource implements StoreDataSource {
     }
 
     @Override
-    public long getCount() {
+    public int getCount() {
         return dao.getCount();
     }
 
     @Override
-    public Maybe<Long> getCountRx() {
+    public Maybe<Integer> getCountRx() {
         return dao.getCountRx();
     }
 
@@ -144,12 +144,12 @@ public class StoreRoomDataSource implements StoreDataSource {
     }
 
     @Override
-    public List<Store> getItems(long limit) {
+    public List<Store> getItems(int limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Store>> getItemsRx(long limit) {
+    public Maybe<List<Store>> getItemsRx(int limit) {
         return null;
     }
 }
